requirements.txt
langchain
python-dotenv
streamlit
openai
youtube-transcript-api
pytube
faiss-cpu
tiktoken
langchain-community
